refactor: 클래스 역할이 명확하게 드로나도록 패키지, 클래스명 변경
