+++ src/main/java/study/racingcar/strategy/ManualMovableStrategy.java
@@ -1,14 +1,14 @@
-package study.racingcar.condition;
+package study.racingcar.strategy;
 
 /**
  * Created by wyparks2@gmail.com on 2019-06-18
  * Blog : http://wyparks2.github.io
  * Github : http://github.com/wyparks2
  */
-public class ManualDeclarationCarMoveCondition implements CarMoveCondition {
+public class ManualMovableStrategy implements MovableStrategy {
     private boolean canMove;
 
-    public ManualDeclarationCarMoveCondition(boolean canMove) {
+    public ManualMovableStrategy(boolean canMove) {
         this.canMove = canMove;
     }
