+++ src/main/java/study/racingcar/strategy/MovableStrategy.java
@@ -1,10 +1,10 @@
-package study.racingcar.condition;
+package study.racingcar.strategy;
 
 /**
  * Created by wyparks2@gmail.com on 2019-06-18
  * Blog : http://wyparks2.github.io
  * Github : http://github.com/wyparks2
  */
-public interface CarMoveCondition {
+public interface MovableStrategy {
     boolean canMove();
 }
